:recycle: [SMS] Migrate
